Added basic floating-point arithmetic execution in interpreted RISC-V cpu

+++ src/Archs/Mips/Zarem.Mips.Emulator/TrapHandlers/MipsTrapContext.cs
@@ -1,8 +1,8 @@
 ﻿// Avishai Dernis 2026
 
 using Zarem.Emulator.Machine;
+using Zarem.Emulator.Machine.CPU;
 using Zarem.Emulator.Machine.Enums;
-using Zarem.Emulator.Machine.Interfaces;
 using Zarem.Emulator.TrapHandlers.Interfaces;
 using Zarem.Mips.Models.Instructions.Enums.Registers;
 
@@ -44,10 +44,10 @@ public readonly struct MipsTrapContext : ITrapContext
     public ulong Argument2 => Cpu[MipsGpRegister.Argument2];
 
     /// <inheritdoc/>
-    public float FloatArgument0 => Cpu.FloatProcessor.Singles[MipsFloatRegister.F12];
+    public float FloatArgument0 => Cpu.FloatProcessor.Singles[(int)MipsFloatRegister.F12];
 
     /// <inheritdoc/>
-    public double DoubleArgument0 => Cpu.FloatProcessor.Doubles[MipsFloatRegister.F12];
+    public double DoubleArgument0 => Cpu.FloatProcessor.Doubles[(int)MipsFloatRegister.F12];
 
     /// <inheritdoc/>
     public ulong Result0
@@ -66,14 +66,14 @@ public readonly struct MipsTrapContext : ITrapContext
     /// <inheritdoc/>
     public float FloatResult0
     {
-        get => Cpu.FloatProcessor.Singles[MipsFloatRegister.F0];
-        set => Cpu.FloatProcessor.Singles[MipsFloatRegister.F0] = value;
+        get => Cpu.FloatProcessor.Singles[(int)MipsFloatRegister.F0];
+        set => Cpu.FloatProcessor.Singles[(int)MipsFloatRegister.F0] = value;
     }
 
     /// <inheritdoc/>
     public double DoubleResult0
     {
-        get => Cpu.FloatProcessor.Doubles[MipsFloatRegister.F0];
-        set => Cpu.FloatProcessor.Doubles[MipsFloatRegister.F0] = value;
+        get => Cpu.FloatProcessor.Doubles[(int)MipsFloatRegister.F0];
+        set => Cpu.FloatProcessor.Doubles[(int)MipsFloatRegister.F0] = value;
     }
 }
